Add ability to customize server engine
